Refactor DOIManager to DOIServiceProvider. Renamed ClientRepository. Updated tests. Added composer script to test with junit output
